# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface CreateApprovalInput {
  symbol: string;
  /** e.g. buy, sell, buy_to_open, sell_to_close */
  action: string;
  /** e.g. stock, option, crypto */
  assetType?: string;
  quantity: number;
  strike?: number;
  expiration?: string;
  estimatedCost: number;
  note?: string;
  /** e.g. algo-signal, manual, webhook */
  source?: string;
}

export type ApprovalStatus = typeof ApprovalStatus[keyof typeof ApprovalStatus];


export const ApprovalStatus = {
  pending_approval: 'pending_approval',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Approval {
  id: string;
  type: string;
  symbol: string;
  name: string;
  quantity: number;
  estimated_price: number;
  estimated_total: number;
  submitted_at: string;
  expires_at: string;
  status: ApprovalStatus;
  reason: string;
  requested_by: string;
  /** @nullable */
  resolved_at?: string | null;
  /** @nullable */
  resolved_note?: string | null;
}

export interface CreateApprovalResponse {
  success: boolean;
  message: string;
  data: Approval;
}

export interface ApprovalCounts {
  pending: number;
  approved: number;
  rejected: number;
}

export interface ApprovalsHistoryResponse {
  success: boolean;
  count: number;
  counts: ApprovalCounts;
  data: Approval[];
}

export interface ApprovalsPendingResponse {
  success: boolean;
  count: number;
  data: Approval[];
}

export interface ApprovalActionInput {
  note?: string;
}

export interface ApprovalActionResponse {
  success: boolean;
  message: string;
  data: Approval;
}

export interface HealthStatus {
  status: string;
  /** @nullable */
  timestamp?: string | null;
  /** @nullable */
  service?: string | null;
}

export interface Portfolio {
  account_number: string;
  total_value: number;
  cash: number;
  invested_value: number;
  day_change: number;
  day_change_percent: number;
  total_return: number;
  total_return_percent: number;
  buying_power: number;
  currency: string;
  updated_at: string;
}

export interface PortfolioResponse {
  success: boolean;
  data: Portfolio;
}

export interface Position {
  id: string;
  symbol: string;
  name: string;
  quantity: number;
  average_buy_price: number;
  current_price: number;
  market_value: number;
  day_change: number;
  day_change_percent: number;
  total_return: number;
  total_return_percent: number;
  equity: number;
  percent_of_portfolio: number;
}

export interface PositionsResponse {
  success: boolean;
  count: number;
  data: Position[];
}

export interface WatchlistItem {
  id: string;
  symbol: string;
  name: string;
  current_price: number;
  day_change: number;
  day_change_percent: number;
  week_52_high: number;
  week_52_low: number;
  market_cap: string;
  volume: number;
  /** @nullable */
  pe_ratio: number | null;
  added_at: string;
}

export interface WatchlistResponse {
  success: boolean;
  count: number;
  data: WatchlistItem[];
}

export interface Quote {
  symbol: string;
  price: number;
  change: number;
  changePercent: number;
  previousClose: number;
  timestamp: string;
  source: string;
}

export interface QuotesResponse {
  success: boolean;
  count: number;
  data: Quote[];
  not_found?: string[];
}

export interface QuoteRefreshInput {
  symbols?: string[];
}

export interface MarketIndex {
  symbol: string;
  name: string;
  price: number;
  change: number;
  changePercent: number;
  /** @nullable */
  ytdChangePercent: number | null;
}

export interface Sector {
  name: string;
  etf: string;
  changePercent: number;
  trend: string;
}

export interface Mover {
  symbol: string;
  changePercent: number;
}

export interface MarketStatus {
  status: string;
  message: string;
  nextEvent: string;
}

export type MarketSummaryDataMovers = {
  gainers: Mover[];
  losers: Mover[];
};

export interface MarketSummaryData {
  market: MarketStatus;
  indices: MarketIndex[];
  sectors: Sector[];
  movers: MarketSummaryDataMovers;
  timestamp: string;
}

export interface MarketSummaryResponse {
  success: boolean;
  data: MarketSummaryData;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  buy: 'buy',
  sell: 'sell',
  dividend: 'dividend',
  deposit: 'deposit',
  withdrawal: 'withdrawal',
} as const;

export type ActivityItemStatus = typeof ActivityItemStatus[keyof typeof ActivityItemStatus];


export const ActivityItemStatus = {
  completed: 'completed',
  pending: 'pending',
  cancelled: 'cancelled',
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  /** @nullable */
  symbol?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  quantity?: number | null;
  /** @nullable */
  price?: number | null;
  amount: number;
  status: ActivityItemStatus;
  date: string;
  description: string;
}

export interface ActivitySummary {
  total_invested: string;
  total_proceeds: string;
  total_dividends: string;
  total_deposits: string;
  pending_count: number;
}

export interface ActivityResponse {
  success: boolean;
  count: number;
  summary: ActivitySummary;
  data: ActivityItem[];
}

export type GetQuotesParams = {
/**
 * Comma-separated list of symbols (e.g. AAPL,TSLA)
 */
symbols: string;
};

export type GetAccountActivityParams = {
type?: GetAccountActivityType;
status?: GetAccountActivityStatus;
limit?: number;
};

export type GetAccountActivityType = typeof GetAccountActivityType[keyof typeof GetAccountActivityType];


export const GetAccountActivityType = {
  buy: 'buy',
  sell: 'sell',
  dividend: 'dividend',
  deposit: 'deposit',
  withdrawal: 'withdrawal',
} as const;

export type GetAccountActivityStatus = typeof GetAccountActivityStatus[keyof typeof GetAccountActivityStatus];


export const GetAccountActivityStatus = {
  completed: 'completed',
  pending: 'pending',
  cancelled: 'cancelled',
} as const;