# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Trading Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /approvals:
    post:
      operationId: createApproval
      tags: [approvals]
      summary: Create a new pending approval request (mock/in-memory, no real trade)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApprovalResponse"
        "400":
          description: Missing required fields
...
  /approvals/{id}/approve:
    post:
      operationId: approveApproval
      tags: [approvals]
      summary: Approve a pending trade approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionInput"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalActionResponse"
        "404":
          description: Not found
        "409":
          description: Already resolved
  /approvals/{id}/reject:
    post:
      operationId: rejectApproval
      tags: [approvals]
      summary: Reject a pending trade approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalActionInput"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalActionResponse"
        "404":
          description: Not found
        "409":
          description: Already resolved
...
components:
  schemas:
    CreateApprovalInput:
      type: object
      required: [symbol, action, quantity, estimatedCost]
      properties:
        symbol:
          type: string
        action:
          type: string
          description: "e.g. buy, sell, buy_to_open, sell_to_close"
        assetType:
          type: string
          description: "e.g. stock, option, crypto"
        quantity:
          type: number
        strike:
          type: number
        expiration:
          type: string
        estimatedCost:
          type: number
        note:
          type: string
        source:
          type: string
          description: "e.g. algo-signal, manual, webhook"
    CreateApprovalResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/Approval"
      required: [success, message, data]
    Approval:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        symbol:
          type: string
        name:
          type: string
        quantity:
          type: number
        estimated_price:
          type: number
        estimated_total:
          type: number
        submitted_at:
          type: string
        expires_at:
          type: string
        status:
          type: string
          enum: [pending_approval, approved, rejected]
        reason:
          type: string
        requested_by:
          type: string
        resolved_at:
          type: ["string", "null"]
        resolved_note:
          type: ["string", "null"]
      required:
        - id
        - type
        - symbol
        - name
        - quantity
        - estimated_price
        - estimated_total
        - submitted_at
        - expires_at
        - status
        - reason
        - requested_by
...
    ApprovalActionInput:
      type: object
      properties:
        note:
          type: string
    ApprovalActionResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: "#/components/schemas/Approval"
      required: [success, message, data]
...