# lib/api-zod/src/generated/types/createApprovalInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApprovalInput {
  symbol: string;
  /** e.g. buy, sell, buy_to_open, sell_to_close */
  action: string;
  /** e.g. stock, option, crypto */
  assetType?: string;
  quantity: number;
  strike?: number;
  expiration?: string;
  estimatedCost: number;
  note?: string;
  /** e.g. algo-signal, manual, webhook */
  source?: string;
}