# lib/api-zod/src/generated/types/createApprovalResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Approval } from './approval';

export interface CreateApprovalResponse {
  success: boolean;
  message: string;
  data: Approval;
}